chore: rename backup wallet screens
